+++ test/helpers/contracts/TokenBurner.sol
@@ -1,9 +1,10 @@
 pragma solidity ^0.4.23;
 
-import "../interfaces/ISecurityToken.sol";
+import "../../../contracts/interfaces/ISecurityToken.sol";
+import "../../../contracts/interfaces/ITokenBurner.sol";
 import "openzeppelin-solidity/contracts/token/ERC20/ERC20.sol";
 
-contract TokenBurner {
+contract TokenBurner is ITokenBurner {
 
     address public securityToken;
 
@@ -11,11 +12,10 @@ contract TokenBurner {
         securityToken = _securityToken;
     }
 
-    function burn(address /* _burner */, uint256 /* _value */) public returns(bool) {
+    function burn(address /* _burner */, uint256 /* _value */) public view returns(bool) {
         require(msg.sender == securityToken);
         // Add the schematics for the burner( token holder) that backing the burning of the securities
         return true;
     }
-
 
 }